+++ examples/flag_and_option_specifications.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# examples/flags_and_options_aliases.py
+# examples/flags_and_options_specifications.py
 
 
 # imports
@@ -17,17 +17,17 @@ info_lines = (
 
     'CLASP.Python examples',
     ':version:',
-    "Illustrates use of CLASP.Python's use of flags, options, and aliases",
+    "Illustrates use of CLASP.Python's use of flags, options, and specifications",
     '',
 )
 
-# Specify aliases, parse, and checking standard flags
+# Specify specifications, parse, and checking standard flags
 
 flag_Debug = clasp.flag('--debug', alias='-d', help='runs in Debug mode');
 option_Verbosity = clasp.option('--verbosity', alias='-v', help='specifies the verbosity', values=[ 'terse', 'quiet', 'silent', 'chatty' ]);
 flag_Chatty = clasp.flag('--verbosity=chatty', alias='-c');
 
-aliases = (
+specifications = (
 
     flag_Debug,
     option_Verbosity,
@@ -37,15 +37,15 @@ aliases = (
     clasp.VersionFlag(),
 )
 
-args = clasp.parse(sys.argv, aliases)
+args = clasp.parse(sys.argv, specifications)
 
 if args.flagIsSpecified('--help'):
 
-    clasp.show_usage(aliases, exit_code=0, version=version, stream=sys.stdout, info_lines = info_lines)
+    clasp.show_usage(specifications, exit_code=0, version=version, stream=sys.stdout, info_lines = info_lines)
 
 if args.flagIsSpecified('--version'):
 
-    clasp.show_version(aliases, exit_code=0, version=version, stream=sys.stdout)
+    clasp.show_version(specifications, exit_code=0, version=version, stream=sys.stdout)
 
 
 # Program-specific processing of flags/options