+++ tests/test_arguments.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 from pyclasp import Arguments
-from pyclasp import alias, option
+from pyclasp import specification, option
 from pyclasp import Flag
 from pyclasp import Option
 import pyclasp as clasp
@@ -123,7 +123,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -153,7 +153,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -166,7 +166,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   2)
         self.assertEqual(flag.given_name        ,   '--flag2')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   2)
         self.assertEqual(flag.given_label       ,   'flag2')
         self.assertEqual(flag.name              ,   '--flag2')
@@ -197,7 +197,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -210,7 +210,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   2)
         self.assertEqual(flag.given_name        ,   '--flag2')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   2)
         self.assertEqual(flag.given_label       ,   'flag2')
         self.assertEqual(flag.name              ,   '--flag2')
@@ -223,7 +223,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   3)
         self.assertEqual(flag.given_name        ,   '---x')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   3)
         self.assertEqual(flag.given_label       ,   'x')
         self.assertEqual(flag.name              ,   '---x')
@@ -257,7 +257,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o1')
-        self.assertEqual(option.argument_alias  ,   None)
+        self.assertIsNone(option.argument_specification)
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o1')
         self.assertEqual(option.name            ,   '-o1')
@@ -289,7 +289,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o1')
-        self.assertEqual(option.argument_alias  ,   None)
+        self.assertIsNone(option.argument_specification)
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o1')
         self.assertEqual(option.name            ,   '-o1')
@@ -303,7 +303,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   2)
         self.assertEqual(option.given_name      ,   '--option2')
-        self.assertEqual(option.argument_alias  ,   None)
+        self.assertIsNone(option.argument_specification)
         self.assertEqual(option.given_hyphens   ,   2)
         self.assertEqual(option.given_label     ,   'option2')
         self.assertEqual(option.name            ,   '--option2')
@@ -335,7 +335,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o1')
-        self.assertEqual(option.argument_alias  ,   None)
+        self.assertIsNone(option.argument_specification)
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o1')
         self.assertEqual(option.name            ,   '-o1')
@@ -349,7 +349,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   2)
         self.assertEqual(option.given_name      ,   '--option2')
-        self.assertEqual(option.argument_alias  ,   None)
+        self.assertIsNone(option.argument_specification)
         self.assertEqual(option.given_hyphens   ,   2)
         self.assertEqual(option.given_label     ,   'option2')
         self.assertEqual(option.name            ,   '--option2')
@@ -363,7 +363,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   3)
         self.assertEqual(option.given_name      ,   '---the-third-option')
-        self.assertEqual(option.argument_alias  ,   None)
+        self.assertIsNone(option.argument_specification)
         self.assertEqual(option.given_hyphens   ,   3)
         self.assertEqual(option.given_label     ,   'the-third-option')
         self.assertEqual(option.name            ,   '---the-third-option')
@@ -392,7 +392,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -409,7 +409,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   3)
         self.assertEqual(option.given_name      ,   '--first-option')
-        self.assertEqual(option.argument_alias  ,   None)
+        self.assertIsNone(option.argument_specification)
         self.assertEqual(option.given_hyphens   ,   2)
         self.assertEqual(option.given_label     ,   'first-option')
         self.assertEqual(option.name            ,   '--first-option')
@@ -441,7 +441,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -476,7 +476,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -497,14 +497,14 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertEqual('--option1=v1', args.values[3])
 
 
-    def test_one_flag_and_one_option_and_one_value_with_empty_aliases(self):
+    def test_one_flag_and_one_option_and_one_value_with_empty_specifications(self):
 
-        aliases_list    =   ( tuple(), list(), None )
+        specifications_list    =   ( tuple(), list(), None )
 
-        for aliases in aliases_list:
+        for specifications in specifications_list:
 
             argv    =   ( 'myprog', '-f1', 'value1', '--first-option=val1' )
-            args    =   clasp.parse(argv, aliases)
+            args    =   clasp.parse(argv, specifications)
 
             self.assertEqual('myprog', args.program_name)
 
@@ -517,7 +517,7 @@ class Arguments_tester_1(unittest.TestCase):
             self.assertIsInstance(flag, ( Flag, ))
             self.assertEqual(flag.given_index       ,   1)
             self.assertEqual(flag.given_name        ,   '-f1')
-            self.assertEqual(flag.argument_alias    ,   None)
+            self.assertIsNone(flag.argument_specification)
             self.assertEqual(flag.given_hyphens     ,   1)
             self.assertEqual(flag.given_label       ,   'f1')
             self.assertEqual(flag.name              ,   '-f1')
@@ -534,7 +534,7 @@ class Arguments_tester_1(unittest.TestCase):
             self.assertIsInstance(option, ( Option, ))
             self.assertEqual(option.given_index     ,   3)
             self.assertEqual(option.given_name      ,   '--first-option')
-            self.assertEqual(option.argument_alias  ,   None)
+            self.assertIsNone(option.argument_specification)
             self.assertEqual(option.given_hyphens   ,   2)
             self.assertEqual(option.given_label     ,   'first-option')
             self.assertEqual(option.name            ,   '--first-option')
@@ -550,14 +550,14 @@ class Arguments_tester_1(unittest.TestCase):
             self.assertEqual('value1', args.values[0])
 
 
-    def test_alias_of_flag_with_one_alias(self):
+    def test_alias_of_flag_with_one_specification(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.flag('--verbose', alias = '-v', extras = { 'x-name': 'v-val' }),
         )
         argv    =   ( 'myprog', '--verbose', '--succinct', 'value', '-v' )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -570,7 +570,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '--verbose')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   2)
         self.assertEqual(flag.given_label       ,   'verbose')
         self.assertEqual(flag.name              ,   '--verbose')
@@ -583,7 +583,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   2)
         self.assertEqual(flag.given_name        ,   '--succinct')
-        self.assertEqual(flag.argument_alias    ,   None)
+        self.assertIsNone(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   2)
         self.assertEqual(flag.given_label       ,   'succinct')
         self.assertEqual(flag.name              ,   '--succinct')
@@ -596,7 +596,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   4)
         self.assertEqual(flag.given_name        ,   '-v')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'v')
         self.assertEqual(flag.name              ,   '--verbose')
@@ -614,14 +614,14 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertEqual('value', args.values[0])
 
 
-    def alias_of_flag_with_two_aliases(self):
+    def alias_of_flag_with_two_specifications(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.flag('--expand', aliases = ( '-x', '--x', ), extras = { 'some-value': ( 'e', 'x', 't', 'r', 'a', 's', ) }),
         )
         argv    =   ( 'myprog', '-f1', 'value1', '-x', '--delete', '--x', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -634,7 +634,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertFalse(flag.argument_alias)
+        self.assertFalse(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -647,7 +647,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   3)
         self.assertEqual(flag.given_name        ,   '-x')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'x')
         self.assertEqual(flag.name              ,   '--expand')
@@ -660,7 +660,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   4)
         self.assertEqual(flag.given_name        ,   '--delete')
-        self.assertFalse(flag.argument_alias)
+        self.assertFalse(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   2)
         self.assertEqual(flag.given_label       ,   'delete')
         self.assertEqual(flag.name              ,   '--delete')
@@ -673,7 +673,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   5)
         self.assertEqual(flag.given_name        ,   '--x')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   2)
         self.assertEqual(flag.given_label       ,   'x')
         self.assertEqual(flag.name              ,   '--expand')
@@ -695,14 +695,14 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertEqual('value1', args.values[0])
 
 
-    def test_alias_of_option_with_one_alias(self):
+    def test_alias_of_option_with_one_specification(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--option', alias = '-o'),
         )
         argv    =   ( 'myprog', '-f1', 'value1', '-o=value2', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -715,7 +715,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertFalse(flag.argument_alias)
+        self.assertFalse(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -732,7 +732,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   3)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -750,12 +750,12 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_alias_of_option_with_separate_value(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--option', alias = '-o'),
         )
         argv    =   ( 'myprog', '-o', 'value-1', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -772,7 +772,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -788,12 +788,12 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_alias_of_option_that_has_default_with_separate_value(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--option', alias = '-o', default_value = 'def-val-1'),
         )
         argv    =   ( 'myprog', '-o', 'value-1', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -810,7 +810,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -826,12 +826,12 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_alias_of_option_that_has_default_with_separate_value_that_resembles_flag(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--option', alias = '-o', default_value = 'def-val-1'),
         )
         argv    =   ( 'myprog', '-o', '-o', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -848,7 +848,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -864,12 +864,12 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_alias_of_option_that_has_default_with_missing_separate_value(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--option', alias = '-o', default_value = 'def-val-1'),
         )
         argv    =   ( 'myprog', '-o', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -886,7 +886,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -902,12 +902,12 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_alias_of_option_with_attached_empty(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--option', alias = '-o', default_value = 'def-val-1'),
         )
         argv    =   ( 'myprog', '-o=', 'value-2', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -924,7 +924,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -942,13 +942,13 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_flag_alias_of_option_with_value(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--verbosity'),
             clasp.flag('--verbosity=high', alias = '-v'),
         )
         argv    =   ( 'myprog', '-v', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -965,7 +965,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   1)
         self.assertEqual(option.given_name      ,   '-v')
-        self.assertEqual(option.argument_alias  ,   aliases[1])
+        self.assertEqual(option.argument_specification  ,   specifications[1])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'v')
         self.assertEqual(option.name            ,   '--verbosity')
@@ -981,12 +981,12 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_alias_of_option_with_value(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.option('--option', alias = '-o', default_value = 'default-value'),
         )
         argv    =   ( 'myprog', '-f1', 'value-1', '-o=', '-o=given-value-1', '--option=given-value-2', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -1000,7 +1000,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-f1')
-        self.assertFalse(flag.argument_alias)
+        self.assertFalse(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'f1')
         self.assertEqual(flag.name              ,   '-f1')
@@ -1017,7 +1017,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   3)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -1031,7 +1031,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   4)
         self.assertEqual(option.given_name      ,   '-o')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   1)
         self.assertEqual(option.given_label     ,   'o')
         self.assertEqual(option.name            ,   '--option')
@@ -1045,7 +1045,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index     ,   5)
         self.assertEqual(option.given_name      ,   '--option')
-        self.assertEqual(option.argument_alias  ,   aliases[0])
+        self.assertEqual(option.argument_specification  ,   specifications[0])
         self.assertEqual(option.given_hyphens   ,   2)
         self.assertEqual(option.given_label     ,   'option')
         self.assertEqual(option.name            ,   '--option')
@@ -1062,7 +1062,7 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_flags_combined(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.flag('--compile', alias = '-c'),
             clasp.flag('--debug', alias = '-d'),
@@ -1070,7 +1070,7 @@ class Arguments_tester_1(unittest.TestCase):
         )
 
         argv    =   ( 'myprog', '-ced', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -1083,7 +1083,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-ced')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'ced')
         self.assertEqual(flag.name              ,   '--compile')
@@ -1096,7 +1096,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-ced')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'ced')
         self.assertEqual(flag.name              ,   '--execute')
@@ -1109,7 +1109,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-ced')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'ced')
         self.assertEqual(flag.name              ,   '--debug')
@@ -1128,7 +1128,7 @@ class Arguments_tester_1(unittest.TestCase):
 
     def test_flags_of_flags_and_options_combined(self):
 
-        aliases =   (
+        specifications =   (
 
             clasp.flag('--compile', alias = '-c'),
             clasp.flag('--mode=debug', alias = '-d'),
@@ -1137,7 +1137,7 @@ class Arguments_tester_1(unittest.TestCase):
         )
 
         argv    =   ( 'myprog', '-ced', )
-        args    =   clasp.parse(argv, aliases)
+        args    =   clasp.parse(argv, specifications)
 
         self.assertEqual('myprog', args.program_name)
 
@@ -1150,7 +1150,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-ced')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'ced')
         self.assertEqual(flag.name              ,   '--compile')
@@ -1163,7 +1163,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(flag, ( Flag, ))
         self.assertEqual(flag.given_index       ,   1)
         self.assertEqual(flag.given_name        ,   '-ced')
-        self.assertTrue(flag.argument_alias)
+        self.assertTrue(flag.argument_specification)
         self.assertEqual(flag.given_hyphens     ,   1)
         self.assertEqual(flag.given_label       ,   'ced')
         self.assertEqual(flag.name              ,   '--execute')
@@ -1180,7 +1180,7 @@ class Arguments_tester_1(unittest.TestCase):
         self.assertIsInstance(option, ( Option, ))
         self.assertEqual(option.given_index       ,   1)
         self.assertEqual(option.given_name        ,   '-ced')
-        self.assertTrue(option.argument_alias)
+        self.assertTrue(option.argument_specification)
         self.assertEqual(option.given_hyphens     ,   1)
         self.assertEqual(option.given_label       ,   'ced')
         self.assertEqual(option.name              ,   '--mode')