+++ tests/test_cli.py
@@ -13,9 +13,9 @@ class Util_tester(unittest.TestCase):
         stm     =   StringIO()
 
         try:
-            aliases =   ()
+            specifications =   ()
 
-            clasp.show_version(aliases, version = "1.2.3", stream = stm, program_name = 'myprog', version_prefix = 'v')
+            clasp.show_version(specifications, version = "1.2.3", stream = stm, program_name = 'myprog', version_prefix = 'v')
 
             self.assertEqual('myprog v1.2.3', stm.getvalue().strip())
 
@@ -28,9 +28,9 @@ class Util_tester(unittest.TestCase):
         stm     =   StringIO()
 
         try:
-            aliases =   ()
+            specifications =   ()
 
-            clasp.show_version(aliases, version = [ 1, 2, 3 ], stream = stm, program_name = 'myprog', version_prefix = 'v')
+            clasp.show_version(specifications, version = [ 1, 2, 3 ], stream = stm, program_name = 'myprog', version_prefix = 'v')
 
             self.assertEqual('myprog v1.2.3', stm.getvalue().strip())
 
@@ -43,9 +43,9 @@ class Util_tester(unittest.TestCase):
         stm     =   StringIO()
 
         try:
-            aliases =   ()
+            specifications =   ()
 
-            clasp.show_version(aliases, version = "1.2.3", stream = stm, program_name = 'myprog')
+            clasp.show_version(specifications, version = "1.2.3", stream = stm, program_name = 'myprog')
 
             self.assertEqual('myprog 1.2.3', stm.getvalue().strip())